+++ src/components/Modal/TransitionsModal/demo/index.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import styled from '@emotion/styled';
+import SimpleModal from '../index';
+
+export default function TransitionModalDemo() {
+  return (
+    <>
+      <SimpleModal></SimpleModal>
+    </>
+  );
+}
