+++ src/components/Modal/TransitionsModal/index.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function TransitionModal() {
+  return <>TransitionModal</>;
+}
